Move helper fns to catalog-utils
